添加人员到指定会议

+++ src/main/java/com/fnst/face/mapper/MeetingUserMapper.java
@@ -18,5 +18,5 @@ public interface MeetingUserMapper {
 
     int updateByPrimaryKey(MeetingUser record);
 
-    List<MeetingUser> selectByMeetingId(Integer meetingId);
+    List<MeetingUser> selectByMeetingId(Long meetingId);
 }
